feat(frontend): add update/reset/delete template
